requirements.txt
streamlit
openai-whisper
ffmpeg-python
fpdf
transformers>=4.30.0
torch
